Refactored to address review comments
